+++ src/components/Index.jsx
@@ -1,5 +1,5 @@
-export { default as Login } from "./Login.jsx";
-export { default as Register } from "./Register.jsx";
+export { default as Login } from "./auth/Login.jsx";
+export { default as Register } from "./auth/Register.jsx";
 export { default as Beginner } from "./Beginner.jsx";
 export { default as MainNav } from "./MainNav.jsx";
 export { default as LandingPage } from "./LandingPage.jsx";
