+++ src/components/auth/Register.jsx
@@ -1,16 +1,15 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
+import MainNav from "../MainNav";
 
-import MainNav from "./MainNav";
-import { loginUser } from "../actions";
-
-class Login extends Component {
+export default class Register extends Component {
   constructor(props) {
     super(props);
     this.state = {
+      username: "",
       email: "",
       password: "",
-      remember: "off"
+      password2: ""
     };
   }
 
@@ -20,14 +19,9 @@ class Login extends Component {
     });
   };
 
-  handleCheck = () => {
-    this.setState({
-      remember: this.state.remember === "on" ? "off" : "on"
-    });
-  };
-
-  loginUser = () => {
-    this.props.loginUser(this.state);
+  handleSubmit = e => {
+    e.preventDefault();
+    console.log("dispatching");
   };
 
   render() {
@@ -38,9 +32,9 @@ class Login extends Component {
           <div className="hero-body">
             <div className="container has-text-centered">
               <div className="column is-4 is-offset-4">
-                <h3 className="title has-text-grey">Login</h3>
+                <h3 className="title has-text-grey">Register</h3>
                 <p className="subtitle has-text-grey">
-                  Please login to proceed.
+                  Please register to proceed.
                 </p>
                 <div className="box">
                   <figure className="avatar">
@@ -50,15 +44,27 @@ class Login extends Component {
                       src="/images/avatar.png"
                     />
                   </figure>
-                  <form>
+                  <form onSubmit={this.handleSubmit}>
+                    <div className="field">
+                      <div className="control">
+                        <input
+                          className="input is-large"
+                          type="text"
+                          placeholder="Username"
+                          autoFocus=""
+                          name="username"
+                          onChange={this.handleInputChange}
+                        />
+                      </div>
+                    </div>
                     <div className="field">
                       <div className="control">
                         <input
                           className="input is-large"
                           type="email"
-                          name="email"
-                          placeholder="Your Email"
+                          placeholder="Email"
                           autoFocus=""
+                          name="email"
                           onChange={this.handleInputChange}
                         />
                       </div>
@@ -69,33 +75,31 @@ class Login extends Component {
                         <input
                           className="input is-large"
                           type="password"
+                          placeholder="Password"
                           name="password"
-                          placeholder="Your Password"
                           onChange={this.handleInputChange}
                         />
                       </div>
                     </div>
                     <div className="field">
-                      <label className="checkbox">
+                      <div className="control">
                         <input
-                          type="checkbox"
-                          name="remember"
-                          onClick={this.handleCheck}
+                          className="input is-large"
+                          type="password"
+                          placeholder="Confirm Password"
+                          name="password2"
+                          onChange={this.handleInputChange}
                         />
-                        Remember me
-                      </label>
+                      </div>
                     </div>
-                    <button
-                      onClick={this.loginUser}
-                      className="button is-block is-info is-large is-fullwidth"
-                    >
-                      Login
+
+                    <button className="button is-block is-info is-large is-fullwidth">
+                      Register
                     </button>
                   </form>
                 </div>
                 <p className="has-text-grey">
-                  <a href="#/register">Sign Up</a> &nbsp;·&nbsp;
-                  <a href="#/forgot">Forgot Password</a> &nbsp;·&nbsp;
+                  <a href="#/login">Login</a> &nbsp;·&nbsp;
                   <a href="#/help">Need Help?</a>
                 </p>
               </div>
@@ -106,16 +110,3 @@ class Login extends Component {
     );
   }
 }
-
-const mapDispatchToProps = dispatch => {
-  return {
-    loginUser: creds => {
-      return dispatch(loginUser(creds));
-    }
-  };
-};
-
-export default connect(
-  null,
-  mapDispatchToProps
-)(Login);
